Move fork recovery tests to sui-e2e-tests with deterministic path-exclusive forking
